// src/app/shared/sdk/rest-api/model/jsonReportSuggestionsResponse.ts
/**
 * CyberReportHub API
 * The API to interact with CyberReportHub
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { JsonArticleReportResponse } from './jsonArticleReportResponse';
import { JsonStatsResponse } from './jsonStatsResponse';

export interface JsonReportSuggestionsResponse {
  articles: Array<JsonArticleReportResponse>;
  stats: Array<JsonStatsResponse>;
}